refactor(eval): extract provenance helpers into evolution/training/_provenance.py
Honors "Never duplicate content across files" — git_sha, git_dirty, and
sha256_file are now shared between the step-1 dataset builder and the
upcoming step-2 training driver. Zero behavior change.

The dataset builder loses its hashlib/subprocess imports (only used by the
extracted helpers) and gains a single import line from _provenance.
